Stop timer when buffer is killed
This makes it so the user doesn't have to manually call identica-stop
before killing the buffer, it also fixes a bug where the timer would
set it's mode name in different buffers and happily continue.

Steps to reproduce: Run identica-mode, kill it's buffer with C-x k,
next time the timer action runs it will set the current buffer's mode name.