Fix pointer moving upon background dent loading. patch by Alexandre Oliva
The problem is that, when some other buffer is active, including the
minibuffer, we switch the wrong window to the identica-buffer, clear
the buffer (moving the point on the right window), then fill it in and
move the point on the wrong window.

The fix is to only clear and re-create the identica-buffer when it is
the active buffer, and defer updates otherwise.  Thus, if we're on
another window, including the minibuffer, we'll change the status
modeline to indicate there are pending dents.

If there are pending dents, process them instead of downloading more.